feat : GH-47 Add people details module
