+++ feature/peopleDetails/src/main/java/com/jerryokafor/feature/peopledetails/ui/PeopleDetailsScreen.kt
@@ -24,14 +24,10 @@
 
 package com.jerryokafor.feature.peopledetails.ui
 
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.tween
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.FlowRow
@@ -43,7 +39,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
@@ -55,32 +50,18 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.contentColorFor
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawWithCache
 import androidx.compose.ui.draw.scale
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.graphics.drawscope.clipRect
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.drawText
-import androidx.compose.ui.text.rememberTextMeasurer
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.Constraints
-import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.jerryokafor.feature.peopledetails.R
@@ -88,7 +69,6 @@ import com.jerryokafor.feature.peopledetails.viewModel.PeopleDetailsViewModel
 import com.jerryokafor.feature.peopledetails.viewModel.PersonDetailsUiState
 import com.valentinilk.shimmer.ShimmerBounds
 import com.valentinilk.shimmer.rememberShimmer
-import kotlinx.coroutines.launch
 import me.jerryokafor.core.common.annotation.ExcludeFromGeneratedCoverageReport
 import me.jerryokafor.core.data.util.ImageUtil
 import me.jerryokafor.core.ds.annotation.ThemePreviews
@@ -101,6 +81,7 @@ import me.jerryokafor.core.ds.theme.ThreeVerticalSpacer
 import me.jerryokafor.core.ds.theme.TwoVerticalSpacer
 import me.jerryokafor.core.model.Credit
 import me.jerryokafor.core.model.PersonDetails
+import me.jerryokafor.core.model.Timeline
 import me.jerryokafor.core.ui.components.Background
 import me.jerryokafor.core.ui.components.IheNkiriCircularProgressIndicator
 import me.jerryokafor.core.ui.components.IhenkiriCollapsingToolbarHeader
@@ -118,32 +99,6 @@ fun PeopleDetailsScreenPreview() {
         Background {
             PeopleDetailsScreen(uiState = PersonDetailsUiState.Loading)
         }
-    }
-}
-
-@Suppress("MagicNumber")
-private val timelines = (0..6).mapIndexed { index, item ->
-    object : TimelineScope {
-        override val title: String
-            get() = "202$index"
-        override val content: @Composable (BoxScope.(Modifier) -> Unit)
-            get() = {
-                Surface(
-                    modifier = it.padding(top = 16.dp, start = 16.dp),
-                    onClick = {},
-                ) {
-                    Column {
-                        Text(
-                            modifier = Modifier,
-                            text = "This is a timeline",
-                        )
-                        Text(
-                            modifier = Modifier.padding(start = 32.dp),
-                            text = "as Tiler",
-                        )
-                    }
-                }
-            }
     }
 }
 
@@ -222,18 +177,21 @@ fun PeopleDetailsScreen(
                                     text = uiState.personDetails.biography,
                                     style = IheNkiri.typography.bodyLarge,
                                     textAlign = TextAlign.Justify,
+                                    overflow = TextOverflow.Ellipsis,
+                                    maxLines = 11,
                                     color = contentColorFor(IheNkiri.color.inverseOnSurface)
                                         .copy(alpha = 0.75f),
                                 )
                             }
 
                             // Known for
-                            KnownForRow(credits = uiState.personDetails.knownFor)
+                            KnownForRow(
+                                credits = uiState.personDetails.knownFor.filter { it.title != null },
+                            )
                             TwoVerticalSpacer()
 
                             Column(
-                                modifier = Modifier
-                                    .padding(horizontal = IheNkiri.spacing.two),
+                                modifier = Modifier.padding(horizontal = IheNkiri.spacing.two),
                             ) {
                                 Text(
                                     text = stringResource(R.string.acting_timeline),
@@ -243,6 +201,9 @@ fun PeopleDetailsScreen(
                                     ),
                                 )
                                 OneAndHalfVerticalSpacer()
+                                val timelines = uiState.personDetails.timeline
+                                    .map { (key, value) -> PersonTimelineView(key, value) }
+
                                 Timeline(modifier = Modifier, items = timelines)
                             }
                         }
@@ -270,13 +231,13 @@ fun PeopleDetailsScreen(
 }
 
 @Composable
-fun PersonalInfo(personDetails: PersonDetails) {
+private fun PersonalInfo(personDetails: PersonDetails) {
     Column(
         modifier = Modifier.padding(horizontal = IheNkiri.spacing.two),
         verticalArrangement = Arrangement.Center,
     ) {
         Text(
-            text = "Personal Info",
+            text = stringResource(R.string.personal_info),
             style = IheNkiri.typography.titleLarge,
             color = contentColorFor(
                 backgroundColor = IheNkiri.color.inverseOnSurface,
@@ -403,7 +364,7 @@ fun PersonalInfo(personDetails: PersonDetails) {
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
-fun SocialButtonRow() {
+private fun SocialButtonRow() {
     FlowRow(
         modifier = Modifier
             .padding(horizontal = IheNkiri.spacing.two)
@@ -414,7 +375,7 @@ fun SocialButtonRow() {
             Icon(
                 modifier = @Suppress("MagicNumber") Modifier.scale(0.5f),
                 painter = painterResource(id = me.jerryokafor.core.ui.R.drawable.ic_facebook),
-                contentDescription = "Facebook",
+                contentDescription = stringResource(R.string.cont_desc_facebook),
             )
         }
         Spacer(modifier = Modifier.width(IheNkiri.spacing.oneAndHalf))
@@ -422,7 +383,7 @@ fun SocialButtonRow() {
             Icon(
                 modifier = @Suppress("MagicNumber") Modifier.scale(0.5f),
                 painter = painterResource(id = me.jerryokafor.core.ui.R.drawable.ic_twitter),
-                contentDescription = "Facebook",
+                contentDescription = stringResource(R.string.cont_desc_twitter),
             )
         }
         Spacer(modifier = Modifier.width(IheNkiri.spacing.oneAndHalf))
@@ -430,7 +391,7 @@ fun SocialButtonRow() {
             Icon(
                 modifier = @Suppress("MagicNumber") Modifier.scale(0.5f),
                 painter = painterResource(id = me.jerryokafor.core.ui.R.drawable.ic_instagram),
-                contentDescription = "Facebook",
+                contentDescription = stringResource(R.string.cont_desc_instagram),
             )
         }
         Spacer(modifier = Modifier.width(IheNkiri.spacing.oneAndHalf))
@@ -438,14 +399,14 @@ fun SocialButtonRow() {
             Icon(
                 modifier = @Suppress("MagicNumber") Modifier.scale(0.5f),
                 painter = painterResource(id = me.jerryokafor.core.ui.R.drawable.ic_tiktok),
-                contentDescription = "Facebook",
+                contentDescription = stringResource(R.string.cont_desc_tiktok),
             )
         }
     }
 }
 
 @Composable
-fun KnownForRow(credits: List<Credit>) {
+private fun KnownForRow(credits: List<Credit>) {
     val shimmer = rememberShimmer(shimmerBounds = ShimmerBounds.Window)
 
     Column(modifier = Modifier) {
@@ -472,161 +433,30 @@ fun KnownForRow(credits: List<Credit>) {
                     size = ImageUtil.Size.Poster.W154,
                 )
                 @Suppress("MagicNumber")
-                MoviePoster(
-                    modifier = Modifier
-                        .width(120.dp)
-                        .aspectRatio(0.7F),
-                    posterUrl = path,
-                    contentDescription = it.title ?: "",
-                    shimmer = shimmer,
-                    onClick = { },
-                )
-            }
-        }
-    }
-}
-
-data class TimelineNodeConfig(
-    val indicatorColor: Color,
-    val circleRadius: Dp,
-    val circleStrokeWidth: Dp,
-    val lineStrokeWidth: Dp,
-    val titleTextStyle: TextStyle,
-)
-
-object TimelineNodeDefaults {
-    val indicatorColor: Color = Color.White
-    val circleRadius: Dp = 8.dp
-    val circleStrokeWidth: Dp = 3.0.dp
-    val lineStrokeWidth: Dp = 1.0.dp
-    val titleTextStyle: TextStyle = TextStyle(fontSize = 14.sp)
-
-    @Composable
-    fun timelineConfig(
-        indicatorColor: Color = TimelineNodeDefaults.indicatorColor,
-        circleRadius: Dp = TimelineNodeDefaults.circleRadius,
-        circleStrokeWidth: Dp = TimelineNodeDefaults.circleStrokeWidth,
-        lineStrokeWidth: Dp = TimelineNodeDefaults.lineStrokeWidth,
-        titleTextStyle: TextStyle = TimelineNodeDefaults.titleTextStyle,
-    ): TimelineNodeConfig = TimelineNodeConfig(
-        indicatorColor = indicatorColor,
-        circleRadius = circleRadius,
-        circleStrokeWidth = circleStrokeWidth,
-        lineStrokeWidth = lineStrokeWidth,
-        titleTextStyle = titleTextStyle,
-    )
-}
-
-interface TimelineScope {
-    val title: String
-    val content: @Composable BoxScope.(Modifier) -> Unit
-}
-
-@Composable
-fun <T : TimelineScope> Timeline(
-    modifier: Modifier = Modifier,
-    items: List<T>,
-) {
-    val lineAnim = remember { Animatable(0f) }
-    val circleAnim = remember { Animatable(0f) }
-
-    Column(modifier = modifier, verticalArrangement = Arrangement.Center) {
-        items.forEach { item ->
-            TimelineNode(
-                title = item.title,
-                lineAnim = lineAnim.value,
-                circleAnim = circleAnim.value,
-            ) {
-                item.content(this, it)
-            }
-        }
-    }
-
-    LaunchedEffect(Unit, block = {
-        launch {
-            lineAnim.animateTo(
-                targetValue = 1f,
-                animationSpec = @Suppress("MagicNumber")
-                tween(2000, easing = LinearEasing),
-            )
-        }
-        launch {
-            circleAnim.animateTo(
-                targetValue = 1f,
-                animationSpec = @Suppress("MagicNumber")
-                tween(2000, easing = LinearEasing),
-            )
-        }
-    })
-}
-
-@Composable
-fun TimelineNode(
-    modifier: Modifier = Modifier,
-    title: String,
-    lineAnim: Float,
-    circleAnim: Float,
-    config: TimelineNodeConfig = TimelineNodeDefaults.timelineConfig(),
-    content: @Composable BoxScope.(Modifier) -> Unit,
-) {
-    val textMeasurer = rememberTextMeasurer()
-    val linePadding = 8.dp
-
-    Box(
-        modifier = modifier
-            .wrapContentSize()
-            .padding(start = 8.dp, top = 8.dp)
-            .drawWithCache {
-                val textLayoutResult = textMeasurer.measure(
-                    text = AnnotatedString(title),
-                    constraints = Constraints.fixedWidth(size.width.toInt()),
-                    style = config.titleTextStyle,
-                )
-
-                onDrawBehind {
-                    val circleRadiusPx = config.circleRadius.toPx()
-                    val linePaddingPx = linePadding.toPx()
-                    val circleStrokeSizePx = config.circleStrokeWidth.toPx()
-                    val lineStrokeSizePx = config.lineStrokeWidth.toPx()
-                    val timelineColors = listOf(Color.Cyan, Color.LightGray)
-
-                    drawArc(
-                        color = config.indicatorColor,
-                        startAngle = 270f,
-                        sweepAngle =
-                            @Suppress("MagicNumber")
-                            360f
-                                *
-                                circleAnim,
-                        useCenter = false,
-                        topLeft = Offset(x = 0f, y = 0f),
-                        size = Size(width = circleRadiusPx * 2, height = circleRadiusPx * 2),
-                        style = Stroke(width = circleStrokeSizePx),
+                Column(
+                    modifier = Modifier.width(120.dp),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                ) {
+                    MoviePoster(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .aspectRatio(0.7F),
+                        posterUrl = path,
+                        contentDescription = it.title ?: "",
+                        shimmer = shimmer,
+                        onClick = { },
                     )
-
-                    drawText(
-                        topLeft = Offset(circleRadiusPx * 2 + linePaddingPx, 0F),
-                        textLayoutResult = textLayoutResult,
-                        color = Color.White,
+                    OneAndHalfVerticalSpacer()
+                    Text(
+                        text = it.title ?: "",
+                        style = IheNkiri.typography.bodySmall,
+                        textAlign = TextAlign.Center,
+                        minLines = 3,
+                        maxLines = 3,
+                        overflow = TextOverflow.Ellipsis,
                     )
-
-                    clipRect(bottom = size.height * lineAnim) {
-                        drawLine(
-                            brush = Brush.verticalGradient(timelineColors),
-                            start = Offset(
-                                x = circleRadiusPx,
-                                y = circleRadiusPx * 2 + linePaddingPx,
-                            ),
-                            end = Offset(
-                                x = circleRadiusPx,
-                                y = size.height,
-                            ),
-                            cap = StrokeCap.Round,
-                            strokeWidth = lineStrokeSizePx,
-                        )
-                    }
                 }
-            },
-        content = { content(Modifier) },
-    )
+            }
+        }
+    }
 }
