+++ feature/peopleDetails/src/main/java/com/jerryokafor/feature/peopledetails/ui/PersonTimelineView.kt
@@ -1,0 +1,70 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2023 IheNkiri Project
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package com.jerryokafor.feature.peopledetails.ui
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import me.jerryokafor.core.ds.theme.IheNkiri
+import me.jerryokafor.core.model.Timeline
+
+class PersonTimelineView(private val key: Int, private val value: List<Timeline>) : TimelineScope {
+    override val title: String
+        get() = "$key"
+    override val content: @Composable (BoxScope.(Modifier) -> Unit)
+        get() = {
+            Box(
+                modifier = it.padding(
+                    top = IheNkiri.spacing.two,
+                    start = IheNkiri.spacing.two,
+                ),
+            ) {
+                Column(verticalArrangement = Arrangement.spacedBy(IheNkiri.spacing.one)) {
+                    value.forEach {
+                        Surface(onClick = {}, shape = IheNkiri.shape.small) {
+                            Column(modifier = Modifier.padding(IheNkiri.spacing.one)) {
+                                Text(
+                                    modifier = Modifier,
+                                    style = IheNkiri.typography.titleSmall,
+                                    text = it.title,
+                                )
+                                Text(
+                                    modifier = Modifier.padding(start = IheNkiri.spacing.four),
+                                    style = IheNkiri.typography.bodyMedium,
+                                    text = it.description,
+                                )
+                            }
+                        }
+                    }
+                }
+            }
+        }
+}
